# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ConversationQueryRequest": {
        "properties": {
          "meeting_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only this meeting's transcript.",
            "title": "Meeting Id"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          },
          "retrieve_only": {
            "default": false,
            "description": "Return the matching transcript chunks as evidence without calling the LLM.",
            "title": "Retrieve Only",
            "type": "boolean"
          },
          "since": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only consider activity on or after this date.",
            "title": "Since"
          },
          "until": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only meetings dated before this date (exclusive).",
            "title": "Until"
          }
        },
        "required": [
          "question"
        ],
        "title": "ConversationQueryRequest",
        "type": "object"
      },
...
      "GroupCreateEdit": {
        "properties": {
          "github_repo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repo Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notify": {
            "default": false,
            "title": "Notify",
            "type": "boolean"
          },
          "owner_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner User Id"
          },
          "trello_board_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trello Board Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "GroupCreateEdit",
        "type": "object"
      },
...
      "GroupOut": {
        "properties": {
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "github_connected": {
            "default": false,
            "title": "Github Connected",
            "type": "boolean"
          },
          "github_repo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repo Url"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/GroupMemberOut"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notify": {
            "default": false,
            "title": "Notify",
            "type": "boolean"
          },
          "trello_board_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trello Board Id"
          },
          "trello_connected": {
            "default": false,
            "title": "Trello Connected",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "created",
          "members"
        ],
        "title": "GroupOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestSummary": {
        "description": "Counts are newly added records, so a second run over unchanged data reports\nzeros. `status` is \"running\" when a large repo was handed to a background task;\npoll the ingest status endpoint for completion.",
        "properties": {
          "commits_ingested": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commits Ingested"
          },
          "group_id": {
            "title": "Group Id",
            "type": "integer"
          },
          "issue_comments_ingested": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Comments Ingested"
          },
          "issues_ingested": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issues Ingested"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Poll GET /jobs/{job_id} while status is running.",
            "title": "Job Id"
          },
          "review_comments_ingested": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comments Ingested"
          },
          "status": {
            "default": "completed",
            "enum": [
              "completed",
              "running"
            ],
            "title": "Status",
            "type": "string"
          },
          "trello_actions_ingested": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trello Actions Ingested"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "group_id"
        ],
        "title": "IngestSummary",
        "type": "object"
      },
      "JobAccepted": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "state": {
            "default": "queued",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
...
      "QueryRequest": {
        "properties": {
          "question": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          },
          "since": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only consider activity on or after this date.",
            "title": "Since"
          }
        },
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
...
      "SourceQueryResponse": {
        "description": "An answer plus the evidence it was built from.",
        "properties": {
          "answer": {
            "default": "",
            "title": "Answer",
            "type": "string"
          },
          "complete_window": {
            "default": false,
            "description": "True if `evidence` is every record in the `since` window, not a relevance sample.",
            "title": "Complete Window",
            "type": "boolean"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "source": {
            "enum": [
              "conversation",
              "github",
              "trello"
            ],
            "title": "Source",
            "type": "string"
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectStatsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "truncated_evidence": {
            "default": 0,
            "title": "Truncated Evidence",
            "type": "integer"
          }
        },
        "required": [
          "source",
          "question",
          "evidence"
        ],
        "title": "SourceQueryResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/users/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/groups/{group_id}": {
      "delete": {
        "operationId": "delete_group_groups__group_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Group",
        "tags": [
          "groups"
        ]
      },
      "get": {
        "operationId": "get_group_groups__group_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Group",
        "tags": [
          "groups"
        ]
      },
      "put": {
        "operationId": "update_group_groups__group_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreateEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Group",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/groups/{group_id}/conversation/query": {
      "post": {
        "description": "Answer a question from this group's meeting transcripts, with the extracts used.\n\nWith `since`/`until` the meeting-date range is applied inside the vector store, and a small\nenough window is read in full (chronologically) rather than relevance-sampled. Set\n`retrieve_only` to get the chunks without an LLM answer.",
        "operationId": "query_conversation_groups__group_id__conversation_query_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
            "in": "query",
            "name": "async",
            "required": false,
            "schema": {
              "default": false,
              "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
              "title": "Async",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceQueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Query Conversation",
        "tags": [
          "query"
        ]
      }
    },
    "/groups/{group_id}/github/query": {
      "post": {
        "description": "Answer a question from the group's ingested GitHub commits, issues and review comments.",
        "operationId": "query_github_groups__group_id__github_query_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
            "in": "query",
            "name": "async",
            "required": false,
            "schema": {
              "default": false,
              "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
              "title": "Async",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceQueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Query Github",
        "tags": [
          "query"
        ]
      }
    },
    "/groups/{group_id}/ingest": {
      "post": {
        "description": "Ingest the group's linked GitHub repo and/or Trello board. Idempotent unless `flush` is set.\nLarge repos are ingested in the background; poll `GET /groups/{group_id}/ingest/status`.",
        "operationId": "ingest_groups__group_id__ingest_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sources",
            "required": false,
            "schema": {
              "default": [
                "github",
                "trello"
              ],
              "items": {
                "enum": [
                  "github",
                  "trello"
                ],
                "type": "string"
              },
              "title": "Sources",
              "type": "array"
            }
          },
          {
            "description": "Discard the selected sources' stored data and rebuild from scratch.",
            "in": "query",
            "name": "flush",
            "required": false,
            "schema": {
              "default": false,
              "description": "Discard the selected sources' stored data and rebuild from scratch.",
              "title": "Flush",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Ingest",
        "tags": [
          "query"
        ]
      }
    },
    "/groups/{group_id}/ingest/status": {
      "get": {
        "operationId": "ingest_status_groups__group_id__ingest_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Ingest Status",
        "tags": [
          "query"
        ]
      }
    },
...
    "/groups/{group_id}/query": {
      "post": {
        "description": "Query across sources. Without `sources`, the question is routed to the sources it needs\nand, if several are used, their answers are merged into one.",
        "operationId": "query_unified_groups__group_id__query_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
            "in": "query",
            "name": "async",
            "required": false,
            "schema": {
              "default": false,
              "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
              "title": "Async",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnifiedQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedQueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Query Unified",
        "tags": [
          "query"
        ]
      }
    },
...
    "/groups/{group_id}/stats": {
      "get": {
        "description": "Facts computed from ingested GitHub/Trello data, with no LLM involved.",
        "operationId": "group_stats_groups__group_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "weeks",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 104,
              "minimum": 1,
              "title": "Weeks",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Group Stats",
        "tags": [
          "query"
        ]
      }
    },
...
    "/groups/{group_id}/trello/query": {
      "post": {
        "description": "Answer a question from the group's ingested Trello card activity.",
        "operationId": "query_trello_groups__group_id__trello_query_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
            "in": "query",
            "name": "async",
            "required": false,
            "schema": {
              "default": false,
              "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
              "title": "Async",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceQueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Query Trello",
        "tags": [
          "query"
        ]
      }
    },
...
  }
}